+++ faststream/specification/asyncapi/v2_6_0/facade.py
