+++ faststream/specification/asyncapi/v3_0_0/__init__.py
@@ -1,4 +1,4 @@
-from .asyncapi import AsyncAPI3
+from .facade import AsyncAPI3
 from .generate import get_app_schema
 
 __all__ = (
