+++ faststream/specification/asyncapi/v3_0_0/facade.py
